Add doc/tags to .gitignore
Why you made the change:

I did this so that generated help tags wouldn't interfere with with this
plugin.

How the change addresses the need:

Ignoring the generated doc/tags file prevents installations that suck the
plugin in as a git submodule and then claim it as dirty after the
doc/tags file is generated. This causes problems with people managing
their .vim setup in git as well as prevents them from upgrading in some
cases.